requirements.txt
redis==5.2.0
cassandra-driver==3.29.2
